feat: Enhance authorization and user profile management
- Refactor authorization middleware to store credentials in request object.
- Update payment intent creation to fetch total amount from order items.
- Modify cart item editing to include item ID and improve error handling.
- Implement user profile deletion with permission checks for admin and client roles.
- Enhance user profile editing service to restrict client users from editing others' profiles.
- Normalize user profile response based on user type (admin/client).
- Introduce role-based access control middleware for route protection.
- Update routes to reflect new authorization structure and improve clarity.
- Adjust Prisma repositories to include necessary user details in responses.
- Fix various error handling and validation issues across services.

+++ src/Controllers/Users/delete-users-profile.controller.ts
@@ -3,6 +3,7 @@ import { UsersDeleteProfileService } from "../../Services/Users/delete-user-prof
 import { prismaService } from "../../lib/prisma.service";
 import { PrismaUsersRepositories } from "../../Repositories/Users/Prisma/PrismaUsersRepositories";
 import { PrismaAccountRepositories } from "../../Repositories/General/Accounts/Prisma/PrismaAccountsRepositories";
+import { RequestWithCredentials } from "../../Common/Middlewares/Authorization/authorization";
 
 const repository: PrismaUsersRepositories = new PrismaUsersRepositories(prismaService)
 const accountRepository: PrismaAccountRepositories = new PrismaAccountRepositories(prismaService)
@@ -10,12 +11,21 @@ const service: UsersDeleteProfileService = new UsersDeleteProfileService(prismaS
 
 class UsersDeleteProfileController
 {
-    static async delete(req: Request, res: Response):Promise<Response | any>
+    static async delete(req: RequestWithCredentials, res: Response):Promise<Response | any>
     {
         try
         {
             const id_user = Number(req.params.id_user)
-            const result = await service.deleteProfile(id_user)
+            const authUser = req.credentials;
+            if (!authUser)
+            {
+                return res.status(401).json({success: false, statusCode:401, message: "Perfil não autenticado." });
+            }
+            if (authUser.user_type === "client" && id_user !== authUser.sub)
+            {
+                return res.status(403).json({success: false, statusCode:403,message: "Você só pode deletar o seu próprio perfil"});
+            }
+            const result = await service.deleteProfile(id_user,{sub: authUser.sub, user_type: authUser.user_type as "admin" | "client"})
             return res.status(result.statusCode).json(result)
         }
         catch (error: any)
